+++ .gitignore
@@ -1,0 +1,133 @@
+.DS_Store
+*.d
+*.o
+*.bin
+*.elf
+*.hex
+*.lss
+*.map
+*.sym
+*.eep
+*.exe
+*.json
+*.bat
+!ExecutePokeConModified-Extension.bat
+.idea/
+SerialController/Captures/
+SerialController/Template/*
+!SerialController/Template/Samples/
+SerialController/profiles/
+SerialController/Controller_Log/
+SerialController/log/
+SerialController/settings.poke
+SerialController/Commands/PythonCommands/*
+!SerialController/Commands/PythonCommands/Samples/
+!SerialController/Commands/PythonCommands/ImageProcessingOnly/
+SerialController/Commands/PythonCommands/ImageProcessingOnly/*
+!SerialController/Commands/PythonCommands/ImageProcessingOnly/Samples/
+
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+.hypothesis/
+.pytest_cache/
+
+# Translations
+*.mo
+*.pot
+
+# Django stuff:
+*.log
+local_settings.py
+db.sqlite3
+
+# Flask stuff:
+instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+target/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# pyenv
+.python-version
+
+# celery beat schedule file
+celerybeat-schedule
+
+# SageMath parsed files
+*.sage.py
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Spyder project settings
+.spyderproject
+.spyproject
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+/site
+
+# mypy
+.mypy_cache/
