+++ ExecutePokeConModified-Extension.bat
@@ -1,3 +1,4 @@
+python SerialController/PokeConUpdateChecker.py
 cd SerialController
 rem python Window.py --profile dragonite
 python Window.py
