+++ LICENSE
@@ -3,6 +3,7 @@ MIT License
 Copyright (c) 2018 @progmem
 Copyright (c) 2019-2020 KawaSwitch
 Copyright (c) 2021-2023 Moi-poke
+Copyright (c) 2023 Futo030
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
