+++ README.md
@@ -42,6 +42,10 @@
     - clrが使用できなくなったことに対する問題に対応。
   - テキストログ2画面表示
     - 常に表示したい情報と処理ログが混在してしまい見づらいという問題を解決するために導入。
+  - メニューバーにヘルプを追加(ver.0.0.0.0.2)
+    - いろいろと役に立ちそうな項目を追加。
+  - 最新版確認機能(ver.0.0.0.0.2)
+    - pokeconのアップデートを容易にするため導入。
 
 - UI面
   - UI刷新
@@ -72,6 +76,7 @@
 pygame
 paho-mqtt (MQTTを使用する場合のみ)
 windows-capture-device-list (Python 3.10以上で使用する場合のみ)
+gitpython
 ```
 
 以下はMoi-Poke様のMODIFIED版の説明になります。
