+++ SerialController/Commands/PythonCommands/ImageProcessingOnly/Samples/screenshot.py
