+++ SerialController/Template/Samples/egg_found.png
