+++ SerialController/Template/Samples/sample_color_HLS.png
