+++ SerialController/Template/Samples/shiny_mark.png
