+++ requirements.txt
@@ -22,6 +22,7 @@ scipy
 pygame
 paho-mqtt
 windows-capture-device-list
+gitpython
 
 #
 # requirements not mentioned in the documents above:
